split uart in three
