+++ software/controller/lib/hal/uart_base.h
@@ -15,14 +15,14 @@ limitations under the License.
 /******************************************************************
  * Serial ports to GUI and debug interface (virtual port over USB)
  * use uart.
- * [RM] Chapter 38 defines the USART registers.
+ * [RM] Chapter 38 defines the USART function.
  *****************************************************************/
 
 #include <cstdint>
+#include <optional>
 
-#include "circular_buffer.h"
-#include "dma.h"
 #include "gpio.h"
+#include "interrupts.h"
 #include "serial_listeners.h"
 #include "units.h"
 
@@ -86,6 +86,17 @@ class Channel {
 
   virtual void EnableTxInterrupt();
   virtual void EnableTxCompleteInterrupt();
+  void ClearTxCompleteInterrupt();
+  void DisableTxInterrupt();
+
+  uint8_t GetRxReg();
+  void PutTxReg(uint8_t byte);
+
+  volatile uint32_t *GetRxAddress();
+  volatile uint32_t *GetTxAddress();
+
+  bool RxRegNotEmpty();
+  bool TxRegEmpty();
 
   void EnableClock();
   void SetupPins(GPIO::Port port, uint8_t tx_pin, uint8_t rx_pin, std::optional<uint8_t> rts_pin,
@@ -93,69 +104,4 @@ class Channel {
   void SetupUARTRegisters(Frequency cpu_frequency, Frequency baud, bool dma_enable = false);
 };
 
-class SoftChannel : public Channel {
- public:
-  explicit SoftChannel(Base base, uint8_t match_char = 0) : Channel(base, match_char){};
-
-  size_t Read(uint8_t *buffer, size_t length, RxListener *rxl = nullptr) override;
-  size_t Write(uint8_t *buffer, size_t length, TxListener *txl = nullptr) override;
-  // Return the number of bytes currently in the receive buffer and ready to be read.
-  size_t RxFull() const override { return rx_data_.FullCount(); };
-  // Returns the number of free locations in the transmit buffer.
-  size_t TxFree() const override { return tx_data_.FreeCount(); };
-
-  void StopTx() override { tx_data_.Flush(); };
-  void StopRx() override { rx_data_.Flush(); };
-
-  void UARTInterruptHandler();
-
- protected:
-  // circular buffers for transmitting and receiving data
-  CircularBuffer<uint8_t, BufferLength> tx_data_;
-  CircularBuffer<uint8_t, BufferLength> rx_data_;
-};
-
-class DMAChannel : public Channel {
- public:
-  // Create a channel with DMA enabled
-  DMAChannel(Base base, DMA::Base dma, uint8_t match_char = 0);
-
-  void Initialize(GPIO::Port port, uint8_t tx_pin, uint8_t rx_pin, std::optional<uint8_t> rts_pin,
-                  std::optional<uint8_t> cts_pin, GPIO::AlternativeFunction alt_function,
-                  Frequency cpu_frequency, Frequency baud);
-
-  size_t Read(uint8_t *buffer, size_t length, RxListener *rxl = nullptr) override;
-  size_t Write(uint8_t *buffer, size_t length, TxListener *txl = nullptr) override;
-
-  // DMA doesn't have an rx buffer, RxFull doesn't make much sense.
-  size_t RxFull() const override { return 0; };
-  // Returns 0 if transfer is in progress (sending is then impossible) and BufferLength otherwise
-  size_t TxFree() const override { return tx_in_progress_ ? 0 : BufferLength; };
-
-  void StopTx() override { tx_dma_->Disable(); };
-  void StopRx() override { rx_dma_->Disable(); };
-
-  void TxDMAInterruptHandler();
-  void RxDMAInterruptHandler();
-
- private:
-  std::optional<DMA::ChannelControl> tx_dma_{std::nullopt};
-  std::optional<DMA::ChannelControl> rx_dma_{std::nullopt};
-  uint8_t request_{0};  // request number for dma setup
-
-  // Store Tx data in our own buffer, in case caller overwrites its data before DMA has finished
-  // transmitting it all
-  uint8_t tx_data_[BufferLength] = {0};
-
-  bool tx_in_progress_{false};
-  bool rx_in_progress_{false};
-};
-
 }  // namespace UART
-
-#if defined(UART_VIA_DMA)
-extern UART::DMAChannel rpi_uart;
-#else
-extern UART::SoftChannel rpi_uart;
-#endif
-extern UART::SoftChannel debug_uart;
